% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_wrapper.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Multi-class Logistic Regression}
\usage{
LRMultiClass(X, y, beta_init = NULL, numIter = 50, eta = 0.1, lambda = 1)
}
\arguments{
\item{X}{n x p training data, 1st column should be 1s to account for intercept}

\item{y}{a vector of size n of class labels, from 0 to K-1}

\item{beta_init}{(optional) initial starting values of beta for the algorithm, should be p x K matrix 

 @return A list containing:
\describe{
  \item{objective}{Vector of objective function values at each iteration, including initial.}
}}

\item{numIter}{number of FIXED iterations of the algorithm, default value is 50}

\item{eta}{learning rate, default value is 0.1}

\item{lambda}{ridge parameter, default value is 1}
}
\description{
Multi-class Logistic Regression
}
\examples{

# Initialize variables
set.seed(123)
n <- 250 
p <- 3 
K <- 3 

# Generate data
X <- cbind(1, matrix(rnorm(n * (p-1)), n, p-1))

beta_true <- matrix(c(1.5, -1, 0.5, -1, 1, -0.5, 0, 0, 0.5), nrow=p, byrow=TRUE)

softmax <- function(z) exp(z)/rowSums(exp(z))

probs <- softmax(X \%*\% beta_true)

y <- apply(probs, 1, function(pr) sample(0:(K-1), 1, prob=pr))

beta_init <- matrix(0, nrow=p, ncol=K)

# Fit LRMultiClass
res_C <- LRMultiClass(X, y, beta_init, numIter=20, eta=0.5, lambda=0.1)



}
