% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-Means Clustering Algorithm}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{n x p matrix containing n data points to cluster}

\item{K}{integer specifying the number of clusters}

\item{M}{(optional) K x p matrix of cluster centers}

\item{numIter}{number of maximal iterations for the algorithm, the default value is 100}
}
\value{
Returns a vector of length n containing class assignments for each data point.
}
\description{
K-Means Clustering Algorithm
}
\examples{
# Simulated data around two clusters
set.seed(12)
X <- rbind(matrix(rnorm(10 * 5, 10), 10, 5), matrix(rnorm(10 * 5, -10), 10, 5))

# Cluster data around centers
MyKmeans(X, 2, M = rbind(rep(10, 5), rep(-10, 5)))
}
